Updated the README a bit.
Updated the README but the changes are not really noteworthy.
Added a small description of the program and what it does and
started with a section on how to extend the application.

However, as I was writing the description I wasn't happy with the
way it works at the moment, so I decided to first change the way
the program can be extended before continuing the README file.